refactor: share static import parser
