+++ tools/audit-page-shared-chunks.mjs
@@ -1,14 +1,11 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { AUTH_OPTIONS_ONLY_SENTINELS, matchesSentinel } from './bundle-boundary-sentinels.mjs';
+import { parseStaticImports } from './static-import-parser.mjs';
 
 const DEFAULT_DIST_PAGES_DIR = path.resolve(process.cwd(), 'dist/pages');
 const PROFILE_MANAGER_IMPORT = './shared/ProfileManager.js';
 const SHARED_IMPORT_WARNING_LIMIT = 5;
-const STATIC_IMPORT_PATTERN = /^import(?!\s*\()/;
-const BARE_IMPORT_SPECIFIER_PATTERN = /^import\s*['"]([^'"]+)['"]/;
-const FROM_SPECIFIER_PATTERN = /\bfrom\s*['"]([^'"]+)['"]/;
-const STATIC_REEXPORT_PATTERN = /^export\s+(?:\*|(?:type\s+)?\{)/;
 
 function parseArgs(argv) {
   const options = { distPagesDir: DEFAULT_DIST_PAGES_DIR };
@@ -42,85 +39,6 @@ function readNextFlagValue(argv, index, flagName) {
     throw new Error(`必須提供 ${flagName} 的值`);
   }
   return value;
-}
-
-function startsStaticDeclaration(line) {
-  return /^\s*(?:import(?!\s*\()|export)\b/.test(line);
-}
-
-function hasOpenStaticDeclaration(statement) {
-  const openBraces = (statement.match(/\{/g) || []).length;
-  const closeBraces = (statement.match(/\}/g) || []).length;
-  const trimmed = statement.trimEnd();
-  return openBraces > closeBraces || /\bfrom\s*$/.test(trimmed) || /,\s*$/.test(trimmed);
-}
-
-function readStaticImportSpecifier(fragment) {
-  if (!STATIC_IMPORT_PATTERN.test(fragment)) {
-    return null;
-  }
-
-  const bareImportMatch = fragment.match(BARE_IMPORT_SPECIFIER_PATTERN);
-  if (bareImportMatch) {
-    return bareImportMatch[1];
-  }
-
-  const fromImportMatch = fragment.match(FROM_SPECIFIER_PATTERN);
-  return fromImportMatch ? fromImportMatch[1] : null;
-}
-
-function readStaticExportSpecifier(fragment) {
-  if (!STATIC_REEXPORT_PATTERN.test(fragment)) {
-    return null;
-  }
-
-  const fromExportMatch = fragment.match(FROM_SPECIFIER_PATTERN);
-  return fromExportMatch ? fromExportMatch[1] : null;
-}
-
-function readStaticSpecifier(fragment) {
-  return readStaticImportSpecifier(fragment) || readStaticExportSpecifier(fragment);
-}
-
-function extractImportSpecifiers(statement) {
-  return statement
-    .split(';')
-    .map(fragment => readStaticSpecifier(fragment.trim()))
-    .filter(Boolean);
-}
-
-function parseStaticImports(sourceText) {
-  const specifiers = [];
-  let pendingStatement = '';
-
-  for (const line of sourceText.split('\n')) {
-    if (pendingStatement) {
-      pendingStatement += `\n${line}`;
-      const matches = extractImportSpecifiers(pendingStatement);
-      if (matches.length > 0 || !hasOpenStaticDeclaration(pendingStatement)) {
-        specifiers.push(...matches);
-        pendingStatement = '';
-      }
-      continue;
-    }
-
-    if (!startsStaticDeclaration(line)) {
-      continue;
-    }
-
-    const matches = extractImportSpecifiers(line);
-    if (matches.length > 0 || !hasOpenStaticDeclaration(line)) {
-      specifiers.push(...matches);
-    } else {
-      pendingStatement = line;
-    }
-  }
-
-  if (pendingStatement) {
-    specifiers.push(...extractImportSpecifiers(pendingStatement));
-  }
-
-  return specifiers.filter(Boolean);
 }
 
 function listPageEntryFiles(distPagesDir) {
